allure-test1.3

+++ src/main/java/ru/github/com/RenzO102/Calculator.java
@@ -5,7 +5,7 @@ public class Calculator {
         return a + b;
     }
 
-    public long difference(long a, long b) {
+    public int difference(int a, int b) {
         return a - b;
     }
